fix: using async fx's for cleaner interactions
