+++ lib/ui/views/home/home_view.dart
@@ -12,9 +12,7 @@ class HomeView extends StatelessWidget {
     final labels = AppLocalizations.of(context);
 
     return Scaffold(
-      appBar: AppBar(
-        toolbarHeight: 0,
-      ),
+      appBar: AppBar(toolbarHeight: 0),
       body: LayoutBuilder(
           builder: (BuildContext context, BoxConstraints viewportConstraints) {
         return SingleChildScrollView(
